crypto-common: rename `serializable_state` module to `hazmat`
Also removes the toplevel re-exports for the traits, forcing users to go
through the now `pub hazmat` module.

There's a lot of misuse potential with these traits, which are intended
to make it possible to serialize/deserialize the internal state of hash
functions. However previously they were presented side-by-side with
other traits, which made that unclear.

This commit deliberately doesn't make other changes to the file so git
will preserve its history:

    diff --git a/crypto-common/src/serializable_state.rs b/crypto-common/src/hazmat.rs
    similarity index 100%
    rename from crypto-common/src/serializable_state.rs
    rename to crypto-common/src/hazmat.rs

It would be very good for a followup commit to massively expand the
documentation around both the module and the traits, especially to spell
out the various ways they can be misused.
