+++ Cargo.toml
@@ -16,4 +16,6 @@ members = [
 ]
 
 [patch.crates-io]
+crypto-common = { path = "./crypto-common" }
+digest = { path = "./digest" }
 signature = { path = "./signature" }
