+++ crypto-common/src/hazmat.rs
